+++ Nonogram/src/Models/Stages/StageBoard.h
@@ -22,5 +22,5 @@ private:
 	bool isPanning; //Indicates that the user is moving the view
 	bool isSwiping; //Indicates that the user is selecting by swiping
 	bool isUndoingSelection; //Indicates that the user started to swiping on the selected tile and want to unselect it
-	sf::Vector2i startPos;
+	sf::Vector2f startPos;
 };
